[GHC-Upgrade]: Fix breakages with GHC 9.2.8 when building `Shims.hs`
This commit updates the code to be able to build with GHC 9.2.8.
This project hasn't been updated in a while so had quite a few breakages.

This commits fix the breakages using the new "ghc"-as-a-library syntax.
It uses the conditional compile directives to use the updated syntax with
GHC 9.2.8 while maintaining the older syntax with previous versions.

"ghc" as in the library had a lot of internal changes as well as external API
and module export changes between version 8.10 and 9.2.
These changes reflect that fact.
The "ghc-plugins" part also had a few changes to take into account.

The Lot of __GLASGOW_HASKELL__ <= 902 checks ensure that only
compatible syntax gets compiled with GHC 9.2.x

Main breakages that i can remember were:
- "Located" type fixes
- "SrcSpanAnn" type fixes
- ghc the library moved a lot of previous types to using type-families
- makes things compatible with newer syntax and version.
- conditional compiles compatibility and making sure right things compiled.
- more of these that i'm sure i can't remember at the moment.