[GHC-Upgrade]: Fix breakages with GHC 9.2.8 when building main `HasFieldPattern.hs` file
This commit updates the code to be able to build with GHC 9.2.8.
This project hasn't been updated in a while so had quite a few breakages.

This commits fix the breakages using the new "ghc"-as-a-library syntax.

"ghc" as in the library had a lot of internal changes as well as external API
and module export changes between version 8.10 and 9.2.
These changes reflect that fact.
The "ghc-plugins" part also had a few changes to take into account.

Conditional compile directives weren't used here because the plugin mentions
that there'd be only one file with GHC conditional changes, which was the
"Shims.hs" file

The fixes here reflect the changes required to update the main
HasFieldPattern.hs file to update GHC 9.2.8 syntax.